requirements.txt
altgraph==0.17
numpy==1.18.2
opencv-python==4.2.0.32
Pillow==7.1.1
pkg-resources==0.0.0
